Added Assignment 2
From Qn 1,,utilize all knowledge obtained from module 1 to visualize the different parameters, patterns, trends, and relationships. Ensure that each plot is saved as an image and is well annotated(labelled).
QN2 From the question 2,, utilize all the knowledge from module 1 to 4,, to describe the different statistical characteristics in the data and ensure to visualize them. Ensure that the different attributes or data collected per individual is detailed enough to describe them.